requirements.txt
Django==1.7
dj-database-url==0.3.0
dj-static==0.0.6
djangorestframework==2.3.14
django-rest-swagger==0.1.14
gunicorn==19.1.0
static3==0.5.1
psycopg2==2.5.1
django-crispy-forms==1.4.0
django-guardian==1.2.4
django-admin-bootstrapped==1.6.6